cleaned up directory
