--- src/controllers.rs
@@ -1,1 +1,0 @@
-pub mod chat_controller;
