+++ src/controllers/chat_controller.rs
@@ -1,9 +1,9 @@
 use std::sync::Arc;
 
-use crate::socketio::server::{ChatHandler, MessageIn};
-use train_messaging_server::{ChatDAO, Chat};
+use crate::dao::{Chat, ChatDAO};
+use crate::socketio::{ChatHandler, MessageIn};
 
-use socketioxide::extract::{SocketRef, Data};
+use socketioxide::extract::{Data, SocketRef};
 pub struct ChatController {
     // chat_dao: ChatDAO,
     chat_handler: Arc<ChatHandler>,
@@ -19,11 +19,14 @@ impl ChatController {
 
     pub async fn register_chat_handlers(&self, socket: SocketRef) {
         let chat_handler = Arc::clone(&self.chat_handler);
-        
-        socket.on("create-chat", move |socket: SocketRef, Data::<Chat>(data)| async move {
-            let chat_handler = Arc::clone(&chat_handler);
-            chat_handler.handle_create_chat(socket, Data(data)).await;
-        });
+
+        socket.on(
+            "create-chat",
+            move |socket: SocketRef, Data::<Chat>(data)| async move {
+                let chat_handler = Arc::clone(&chat_handler);
+                chat_handler.handle_create_chat(socket, Data(data)).await;
+            },
+        );
 
         let chat_handler = Arc::clone(&self.chat_handler);
         socket.on("join", move |socket: SocketRef, Data::<String>(room)| {
@@ -32,10 +35,12 @@ impl ChatController {
         });
 
         let chat_handler = Arc::clone(&self.chat_handler);
-        socket.on("message", move |socket: SocketRef, Data::<MessageIn>(data)| {
-            let chat_handler = Arc::clone(&chat_handler);
-            chat_handler.handle_message(socket, Data(data));
-        });
+        socket.on(
+            "message",
+            move |socket: SocketRef, Data::<MessageIn>(data)| {
+                let chat_handler = Arc::clone(&chat_handler);
+                chat_handler.handle_message(socket, Data(data));
+            },
+        );
     }
 }
-
