+++ src/controllers/mod.rs
@@ -1,0 +1,2 @@
+mod chat_controller;
+pub use chat_controller::ChatController;
