+++ src/dao/mod.rs
@@ -1,0 +1,7 @@
+pub mod base_dao;
+pub mod chat_dao;
+pub mod message_dao;
+
+pub use base_dao::BaseDAO;
+pub use chat_dao::{Chat, ChatDAO};
+pub use message_dao::{Message, MessageDAO};
