+++ src/models/mod.rs
@@ -1,0 +1,3 @@
+mod session_model;
+
+pub use session_model::Session;
