--- src/socketio.rs
@@ -1,1 +1,0 @@
-pub mod server;
