+++ src/socketio/mod.rs
@@ -1,0 +1,2 @@
+mod chat_server;
+pub use chat_server::{ChatHandler, MessageIn, MessageOut, Server};
